requirements.txt
APScheduler==3.6.3
asgiref==3.3.1
certifi==2020.12.5
chardet==4.0.0
Django==3.1.7
django-filter==2.4.0
django-redis==4.12.1
djangorestframework==3.12.2
drf-ujson2==1.6.1
envparse==0.2.0
idna==2.10
phonenumbers==8.12.22
psycopg2-binary==2.8.6
python-dateutil==2.8.1
python-telegram-bot==13.4.1
pytz==2021.1
redis==3.5.3
requests==2.25.1
six==1.15.0
sqlparse==0.4.1
tornado==6.1
typing==3.7.4.3
tzlocal==2.1
ujson==4.0.2
urllib3==1.26.6
